Fix number of samples per step
Previously, the audio callback function would generate samples until it
was done, ignoring which step it was in. Now the audio callback function
will automatically increment the current step and generate the same
amount of samples for each step.

This also means the SDL window can be updated more often which fixes the
delayed input from before.